refactor(go): Rename goproxy-related files
