+++ lib/datasource/go/releases-goproxy.ts
